requirements.txt
streamlit
pytesseract
Pillow
rapidfuzz
ollama
